Extract OpenableElement mixin for Menu, Dialog, and Popup
