+++ nicegui/elements/menu.py
@@ -1,13 +1,11 @@
-from typing_extensions import Self
-
 from ..defaults import DEFAULT_PROPS, resolve_defaults
 from ..events import ClickEventArguments, Handler
 from .context_menu import ContextMenu
 from .item import Item
-from .mixins.value_element import ValueElement
+from .mixins.openable_element import OpenableElement
 
 
-class Menu(ValueElement[bool]):
+class Menu(OpenableElement):
 
     @resolve_defaults
     def __init__(self, *, value: bool = DEFAULT_PROPS['model-value'] | False) -> None:
@@ -27,24 +25,6 @@ class Menu(ValueElement[bool]):
         self._props.add_warning('touch-position',
                                 'The prop "touch-position" is not supported by `ui.menu`. '
                                 'Use "ui.context_menu()" instead.')
-
-    def _render_markdown(self) -> str:
-        return self._children_to_markdown() if self.value else ''
-
-    def open(self) -> Self:
-        """Open the menu."""
-        self.value = True
-        return self
-
-    def close(self) -> Self:
-        """Close the menu."""
-        self.value = False
-        return self
-
-    def toggle(self) -> Self:
-        """Toggle the menu."""
-        self.value = not self.value
-        return self
 
 
 class MenuItem(Item):
