+++ nicegui/elements/popup.py
@@ -1,10 +1,8 @@
-from typing_extensions import Self
-
 from ..defaults import DEFAULT_PROPS, resolve_defaults
-from .mixins.value_element import ValueElement
+from .mixins.openable_element import OpenableElement
 
 
-class Popup(ValueElement[bool]):
+class Popup(OpenableElement):
 
     @resolve_defaults
     def __init__(self, *, value: bool = DEFAULT_PROPS['model-value'] | False) -> None:
@@ -21,21 +19,3 @@ class Popup(ValueElement[bool]):
         :param value: whether the popup is already opened (default: ``False``)
         """
         super().__init__(tag='q-popup-proxy', value=value)
-
-    def _render_markdown(self) -> str:
-        return self._children_to_markdown() if self.value else ''
-
-    def open(self) -> Self:
-        """Open the popup."""
-        self.value = True
-        return self
-
-    def close(self) -> Self:
-        """Close the popup."""
-        self.value = False
-        return self
-
-    def toggle(self) -> Self:
-        """Toggle the popup."""
-        self.value = not self.value
-        return self
